Exercise 5: Fix dataset splitting in k-NN models.
The test dataset was using the labels of the train dataset, which
completely sabotaged the model from performing sensibly.

Further the visualization of the prediction error then compared the
(low-quality) predictions with the actual labels of the train set, while
it should have compared them with the labels of the test set - again
producing a nonsensical visualization of its own.